+++ SGDK_projects/Mugenesis_Demo [VER.001] [SGDK 211] [GEN] [GAME] [FIGHTING]/inc/mg_config.h
@@ -1,0 +1,5 @@
+/* mg_config.h -- configuracao do runtime mugen2sgdk_forge para o Mugenesis_Demo. */
+#ifndef MG_CONFIG_H
+#define MG_CONFIG_H
+#define MG_EXTERNAL_HUD 1      /* HUD de luta do jogo (src/scenes/fight_hud.c) substitui o HUD de texto */
+#endif