+++ SGDK_projects/Mugenesis_Demo [VER.001] [SGDK 211] [GEN] [GAME] [FIGHTING]/inc/scenes/fight_hud.h
@@ -1,0 +1,5 @@
+#ifndef FIGHT_HUD_H
+#define FIGHT_HUD_H
+void FIGHT_HUD_init(void);
+void FIGHT_HUD_update(void);   /* chamar apos MG_fightRender(); so escreve o que mudou */
+#endif